[Ads] renamed namespace
