+++ MaterialsLib/Adsorption/DensityConst.h
@@ -2,7 +2,7 @@
 
 #include "Adsorption.h"
 
-namespace Ads
+namespace Adsorption
 {
 
 class DensityConst : public Adsorption
