+++ MaterialsLib/Adsorption/DensityCook.h
@@ -2,7 +2,7 @@
 
 #include "Adsorption.h"
 
-namespace Ads
+namespace Adsorption
 {
 
 class DensityCook : public Adsorption
