+++ MaterialsLib/Adsorption/DensityDubinin.cpp
@@ -19,7 +19,7 @@ const double c[] = {
 
 }
 
-namespace Ads
+namespace Adsorption
 {
 
 double DensityDubinin::get_adsorbate_density(const double T_Ads) const
