+++ MaterialsLib/Adsorption/DensityMette.cpp
@@ -16,7 +16,7 @@ const double c[] = {
 };
 }
 
-namespace Ads
+namespace Adsorption
 {
 
 double DensityMette::get_adsorbate_density(const double T_Ads) const
