+++ MaterialsLib/Adsorption/density_inert.h.orig
@@ -2,7 +2,7 @@
 
 #include "adsorption.h"
 
-namespace Ads
+namespace Adsorption
 {
 
 class DensityInert final : public Adsorption
